Added support for payload greater than 100KB
The default "deferred" service provided by GAE has the functionality that if size of the function payload exceeds 100KB, the payload is stored in the data store and then retrieved back when required. This improved version of "deferred" service was missing this helpful functionality for long time. I am adding support for it now.